+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 FEATURES:
 - Added support for `runtime` resources
+- Added support for setting `k8s_namespace` and `ecs_prefix` on `prodvana_release_channel.runtime` definitions